more idiotmatic

+++ src/main.rs
@@ -14,7 +14,7 @@ struct Args {
     players: i32,
 
     #[clap(short, long)]
-    rounds: Option<i16>,
+    rounds: Option<i32>,
 }
 
 fn main() {
@@ -26,20 +26,3 @@ fn main() {
     let output_table = Table::new(players).with(Style::rounded()).to_string();
     print!("{}", output_table)
 }
-
-// fn make_entities(n: i32) -> Vec<entity::Entity> {
-//     let v: Vec<entity::Entity> = (0..n).map(|_x| Entity::new()).collect();
-//     v
-// }
-// let alphabet = Alphabet::alphanumeric();
-
-// // The generator takes a u128 as a seed.
-// let seed = 1234;
-
-// // The length of a generated code. This is really a _minimum_ length; larger numbers
-// // will be converted to longer codes since that's the only way to avoid collisions.
-// let length = 4;
-
-// // A small amount of pre-caching work happens when we create the BlockId instance,
-// // so it's good to re-use the same generator where possible.
-// let generator = BlockId::new(alphabet, seed, length);
